init commit
Currently seeing a weird visual bug where the checkerboard gets glitchy. Could be my intel IGP on the GLES3 renderer and Godot 3.3 RC8's inability to handle that well, or it could be that I messed something up when trying to copy the project and adjust all the directories and dependencies for the sake of a clean repo. This all should work just fine as-is, no weird import settings or anything. This worked totally fine when I was using 3.2.3 stable to create the project.

My biggest fear is just getting good video footage for this shader when I try to promote it later. :P I'll at least get that with the old project where naming stuff and folder structure isn't as critical.